move(VJNP-104): subjectGrid.js constants 폴더 생성
