+++ src/components/SubjectSelection/StyledSubjectGrid.jsx
@@ -1,8 +1,5 @@
 import styled from 'styled-components';
-
-// 다른 반응형과 다른 px값에서 GRID가 바뀌어서 GRID_BREAKPOINT 상수 추가
-export const GRID_BREAKPOINT = { tablet: 868, mobile: 767 };
-export const NUM_OF_SUBJECTS = { pc: 8, other: 6 }; // limit 값 담은 상수
+import { GRID_BREAKPOINT } from '../../constants/subjectGrid';
 
 const StyledSubjectGrid = styled.ol`
   display: grid;
